+++ hooks/src/lib.rs
@@ -1,7 +1,9 @@
 //! # Freya Hooks
 //! A collection of hooks to be used in Freya.
 
+mod rope_editor;
 mod text_editor;
+
 mod use_animation;
 mod use_editable;
 mod use_focus;
@@ -11,7 +13,9 @@ mod use_theme;
 #[cfg(feature = "use_camera")]
 mod use_camera;
 
+pub use rope_editor::*;
 pub use text_editor::*;
+
 pub use use_animation::*;
 pub use use_editable::*;
 pub use use_focus::*;
